Port over sum-of-digits and refactor accordingly
I ported it over, renaming the functions to what they corresponded to in number theory. I also went up the module chain and made it so that the module is imported from the top level.

If this package contains a lot of modules, its going to get annoying having to reference functions deep in nested tables, so I think it would be useful to create a utility function that behaves similar to C++'s using statement. I should examine how to implement such a thing and implement it next.

Next up is to implement the 'using' utility function, and then implement an optional 'aliases' utility flag.