Make complex constructor invokable by simply calling name of class
This was done in the Lua version of this that I'm porting over (also written by myself), and so I'm bringing this behavior as well in case I ever used it in Lua without Fennel. It has been tested via REPL and works.

My last commit message got cut short because I accidentally hit enter, so here is the info I wanted to type - I started porting over the complex class I wrote for vanilla Lua to support artistic projects I was writing utilizing Bitty Engine. I know that I can already use the library I wrote in Fennel since it is already written in Lua, but I want to rewrite it in Fennel so that the projects I eventually write can be homogenous in the language they wind up using. I also simply enjoy porting code sometimes, and since I have really had a need for classes yet in the Fennel programs I've written, this serves as a good chance to get acclimated to the OOP prototype paradigm in Fennel. It feels very Common Lispy.

Next up is to continue porting over more the complex number class.