Port over a few metamethods and a function for complex class
Ported over tostring, eq, len, and get-conjugate. All tested via REPL and work. However, it would be worthwhile to elaborate on the eq metamethod to accomodate comparisons between complex numbers and ordinary numbers in Lua, maybe even bools.

Next up is to expand the eq metamethod to cover other types (which can be in either position if they don't have metamethods themselves for it) or to continue porting over other metamethods and functions for the complex class.