Implement a protected vector dimensionality method in preparation for supporting scalar and vector arithmetic
I coded a single function that returns the dimensionality of a vector if possible, and for scalars and other unsupported types returns nil, so that such a check can be used to determine what is and isn't a vector. This will support the skeletoned num-to-vec function that I left in.

The method has been tested via REPL and works as desired.

Next up is to implement num-to-vec, and then refactor a single vector arithmetic method to accomodate scalar ops and see how it looks before extending to all other arithmetic ops.