Implement vector method that converts scalars to vectors
Very straightforward and simple. It was tested via REPL and works as desired.

Next up is to use the two methods just created to write a method that when give a vector/scalar pair, returns the combo with the scalar converted properly to a vector, handling vector pairs and impossible pairs reasonably.