Implement preparational method that ensures scalars can be used in vectorarithmetic
I now have completed the method that will be used at the beginning of each refactored vector arithmetic operator to enable scalars to be used as one of the operands. It was actually quite simple to implement with destructuring.

It has been tested via REPL and is working exactly as desired.

Next up is to move the methods earlier (since they will need to be implemented before they can be used, or at least declared), and then refactor a single vector arithmetic operator as a test before advancing on to refactoring the others.