Refactor vector addition to allow a potential scalar operand
Utilized the recently implemented methods to refactor a single operation as a trial - addition was chosen. Having it tested it via REPL, it is confirmed to be working properly with any combo and ordering of vectors and scalars.

Next up is to refactor the remaining vector arithmetic operators to allow the same thing.