Start implementing matrix class; implement matrix constructor with initializer function argument
Started the matrix class, and the current constructor I have created has been tested briefly via REPL and is working as desired.

There are implementation decisions that I may want to reflect on at some point - the matrix can be a linear array that uses a formula to simulate a rectangular matrix, or (as I have chosen), can be an array of vectors serving as rows. The latter is convenient and allows me to reuse a previous class I wrote. I should think about whether this is optimal though for what I'm trying to achieve.

Next up is to implement the second constructor, which will accept vector varargs to construct a matrix.