Implement secondary matrix constructor that utilizes vector/array varargs
This secondary constructor was tested at length via REPL, and it works exactly as desired, even on edge cases (like where only one vector is provided).

Next up is to implement the call and index metamethods. I intend for the call metamethod to 'intelligently' invoke the proper constructor from the two, and the index metamethod should index the rows table.