Implement __call metamethod for Matrix class to 'intelligently' select appropriate constructor
As usual, tested via REPL and is working exactly as desired.

Next up is to adjust the __indexing metamethod for instances of the matrix class so that indexing them references the rows table. It should be possible to maintain the existing __index parent referencing functionality on top of this, but I guess I'll see.