Implement matrix method to determine if an object is a matrix, vector, scalar, or other
Read the comment written to the committed file using git diff for more details, but the short of it is that this method is needed to support dispatching to appropriate functions when utilizing overloaded operators.

The method has been tested via REPL and works exactly as desired.

Next up is to start implementing addition via the operator and supporting methods.