Start implementing the matrix addition arithmetic operator overload; consider alternative approaches
I started to implement the general matrix addition method that would 'intelligently' dispatch based on type of the 'other' operand (it can be in either position), and I realized that there might be a more generalized way that may be identically efficient speed-wise but could potentially dramatically reduce duplicate code/logic.

I've detailed my considerations and thoughts at length in a comment above the matrix addition overload that is not yet fully implemented - look at the generated diff to quickly read its contents if desired.

Next up is to consider the alternative approach, and either begin to implement, or implement the supporting functions/methods that would be used in its implementation.